requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gviz-api==1.10.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
jsonslicer==0.1.7
keras-nightly==2.8.0.dev2021120408
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
setuptools-scm==6.3.2
six==1.16.0
tb-nightly==2.8.0a20211204
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.5.0
tensorboard-plugin-wit==1.8.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021120409
tf-nightly==2.8.0.dev20211203
tfa-nightly==0.16.0.dev20211116104445
tomli==1.2.2
tqdm==4.62.3
typeguard==2.13.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
zstandard==0.16.0
